# docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      error:
        type: string
    type: object
...
  models.ResponsePayload:
    properties:
      current_page:
        type: integer
      item_limit:
        type: integer
      items: {}
      next_page:
        type: string
      prev_page:
        type: string
      total_item_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  models.UserGear:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_top_category_id:
        type: integer
      container_link_id:
        type: integer
      container_registration_id:
        type: integer
      gear_category_id:
        type: integer
      gear_height:
        type: integer
      gear_id:
        type: integer
      gear_is_container:
        type: boolean
      gear_length:
        type: integer
      gear_manufacture_id:
        type: integer
      gear_name:
        type: string
      gear_status:
        type: boolean
      gear_top_category_id:
        type: integer
      gear_weight:
        type: integer
      gear_width:
        type: integer
      manufacture_id:
        type: integer
      manufacture_name:
        type: string
      top_category_id:
        type: integer
      top_category_name:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
      user_username:
        type: string
      usergear_gear_id:
        type: integer
      usergear_registration_id:
        type: integer
      usergear_user_id:
        type: integer
    type: object
  models.UserGearLink:
    properties:
      usergear_gear_id:
        type: integer
      usergear_registration_id:
        type: integer
      usergear_user_id:
        type: integer
    type: object
  models.UserGearLinkNoID:
    properties:
      usergear_gear_id:
        type: integer
      usergear_user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@sea-shell.no
    name: API Support
  description: This is the API of GoGear
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoGear API
  version: "1.0"
paths:
...
  /api/v1/usergear/{user}/list:
    get:
      consumes:
      - application/json
      description: Get a list a users gear
      parameters:
      - description: Unique ID of user you want to get the Gear of
        in: path
        name: user
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: top categories
        in: query
        items:
          type: integer
        name: topCategory
        type: array
      - collectionFormat: multi
        description: sub categories
        in: query
        items:
          type: integer
        name: category
        type: array
      - collectionFormat: multi
        description: manufacturers
        in: query
        items:
          type: integer
        name: manufacture
        type: array
      - default: all
        description: show container gear only. valid values are true, false, all
        in: query
        name: container
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.UserGear'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List users gear
      tags:
      - User gear
  /api/v1/usergear/insert:
    put:
      consumes:
      - application/json
      description: Insert user registered gear with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserGearLinkNoID'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Insert user registered gear
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/delete:
    delete:
      consumes:
      - application/json
      description: Delete userGear with corresponding ID value
      parameters:
      - description: Unique ID of userGear you want to update
        in: path
        name: userGear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete userGear with ID
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/get:
    get:
      consumes:
      - application/json
      description: Get user registeredgear spessific to ID
      parameters:
      - description: Unique ID of user registered gear you want to get
        in: path
        name: usergear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.UserGear'
      security:
      - BearerAuth: []
      summary: Get user registered gear with ID
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/update:
    post:
      consumes:
      - application/json
      description: Update user registered gear identified by ID
      parameters:
      - description: Unique ID of user registered gear you want to get
        in: path
        name: usergear
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserGearLink'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Update user registered gear with ID
      tags:
      - User gear
...
securityDefinitions:
  BearerAuth:
    description: Include a server-issued JWT as `Bearer <token>`. Endpoints may require
      either the client or admin audience.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"